uis/pyproject.toml
[tool.poetry]
name = "langchain-playground-uis"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "langchain_playground_uis"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^0.27.2"
langchain = "^0.0.107"
gradio = "^3.20.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
